+++ ticketping/src/api.js
@@ -1,5 +1,5 @@
 import Axios from "axios";
-import { API_HOST } from "./Constants";
+import { API_HOST } from "./constant/Constants";
 
 export const axiosInstance = Axios.create({
   baseURL: API_HOST,
